Wherein Andy moves things around again for no good reason
