+++ src/controller/undo.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::model::app_state::AppState;
+use crate::model::app::AppState;
 use crate::model::mods::ModRecord;
 
 /// Push a point onto the undo stack.
