+++ src/model.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod app_state;
+pub mod app;
 pub mod document;
 pub mod mods;
 pub mod pixels;
