+++ src/view/alert.rs
@@ -18,7 +18,7 @@ use druid::WidgetExt;
 use super::button::Button;
 use super::theme;
 use crate::commands;
-use crate::model::app_state::AppState;
+use crate::model::app::AppState;
 
 pub fn unsaved(parent_pos: druid::Point) -> druid::WindowDesc<AppState> {
     let message = build_message("Do you want to save the changes you made?")
