+++ src/view/palette.rs
@@ -14,7 +14,7 @@
 
 use druid::widget::prelude::*;
 
-use crate::model::app_state::AppState;
+use crate::model::app::AppState;
 use crate::view::theme;
 
 /// A palette that displays available colors. Each value is stored as a u32 representation
