+++ src/view/tool.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 
 use druid::widget::prelude::*;
 
-use crate::model::app_state::AppState;
+use crate::model::app::AppState;
 use crate::model::types::ToolType;
 use crate::view::theme;
